requirements.txt
aiohttp>=3.8
pyjwt>=2.6
bcrypt>=4.0
asyncpg>=0.30
